refactor(get-web-crypto): add a crypto selector function
