+++ lib/create-key.js
@@ -2,12 +2,11 @@
 
 var Buffer = require('buffer/').Buffer
 var randomBytes = require('randombytes')
-var subtle = (global.crypto && global.crypto.subtle) ||
-  (global.msCrypto && window.msCrypto.subtle) ||
-  (global.msrCrypto && window.msrCrypto.subtle) ||
-  require('./helpers/msrcrypto').subtle
+var getCrypto = require('./utils/get-web-crypto')
 
 module.exports = function createKey (password, saltArg) {
+  var subtle = getCrypto().subtle
+
   var passwordBuffer = toBuffer(password, 'utf-8', 'password')
   var digest = 'SHA-256'
   var iterations = 100000
