+++ lib/helpers/decrypt-core.js
@@ -3,10 +3,7 @@
 module.exports = decrypt
 
 var Buffer = require('buffer/').Buffer
-var subtle = (global.crypto && global.crypto.subtle) ||
-  (global.msCrypto && window.msCrypto.subtle) ||
-  (global.msrCrypto && window.msrCrypto.subtle) ||
-  require('./msrcrypto').subtle
+var getCrypto = require('../utils/get-web-crypto')
 
 /**
  * Decrypt data.
@@ -21,6 +18,7 @@ function decrypt (key, iv, data, tag, aad) {
   if (key.length !== 32) {
     return Promise.reject(new TypeError('invalid key size'))
   }
+  var subtle = getCrypto().subtle
 
   return subtle.importKey('raw', key, { name: 'AES-GCM' }, true, ['decrypt'])
