+++ lib/helpers/encrypt-core.js
@@ -4,10 +4,7 @@ module.exports = encrypt
 
 var Buffer = require('buffer/').Buffer
 var Promise = require('lie')
-var subtle = (global.crypto && global.crypto.subtle) ||
-  (global.msCrypto && window.msCrypto.subtle) ||
-  (global.msrCrypto && window.msrCrypto.subtle) ||
-  require('./msrcrypto').subtle
+var getCrypto = require('../utils/get-web-crypto')
 
 /**
  * Encrypt data.
@@ -21,6 +18,7 @@ function encrypt (key, iv, data, aad) {
   if (key.length !== 32) {
     return Promise.reject(new TypeError('invalid key size'))
   }
+  var subtle = getCrypto().subtle
 
   return subtle.importKey('raw', key, { name: 'AES-GCM' }, true, ['encrypt'])
