+++ lib/utils/msrcrypto.js
